Added Sessions for login capability
You can now login using usernames in the database, please use the latest
communit schema.
Updated communit database.
Added connectivity using PHP Sessions.
Site now knows if users are logged in or not and redirects users
accordingly.
Added a Home Page.
Tweaked and added a Josh's Form onto the Home Page.
Created a separate template for the Home page and pages that do not
require to be logged in.
Tweaked templates to show certain buttons based on if the user is logged
in or not (Will be edited to determine if Admin/Guest/User is logged in
and show buttons accordingly).
Cleaned up some code.