Implemented side panel functionality, cleanup/comments/indenting
Half of the side panel now updates as you click on markers. The resident
and resident cell part I need to talk with you guys about the database
before we can move forward with that.

I put commented out code blocks at the very bottom of communitymap.php
because they were just getting in the way. If we still need them they
can be found there.

I commented the crap out of the initialize function and corrected as
much of the indenting as I could.