Community Map updates, database updates, icon updates.
Changed the order in which resident info is displayed.
Added email addresses to sub residents info.
Added a hex color  to the database for each head_resident to be used by
Google Markers.
Added a Misc 1000 character TEXT column to the database for head
residents Misc info.
Added arrays to pull hex color and misc info from the database.
Added the Misc info to the side panel (If the resident has any).
Added a grayscale icon to be colored with a hex value from the database.
Changed the way information is checked for before adding to the info
panel.
Removed the second 'Select a Residence' from the side panel.